requirements.txt
djangorestframework
drf-yasg
Django==2.1.4
lxml==4.3.3
PyMySQL==0.9.3
requests>=2.20.0
rsa==4.0
urllib3>=1.24.2
pytz==2019.1